Test register-instance CLI
