+++ tests/test_20_inventory.py
@@ -1,6 +1,7 @@
 import json
 
 import pytest
+from selenium.webdriver.common.by import By
 from selenium.webdriver.support.ui import Select
 from sh import ErrorReturnCode, temboard
 
@@ -37,6 +38,45 @@ def test_web_register(registered_agent, agent_conf, browser, pg_version, ui_url)
     assert "default" in browser.select(fmt(col=server)).text
     assert port in browser.select(fmt(col=agent)).text
     assert pg_version in browser.select(fmt(col=postgres)).text
+
+
+@pytest.fixture(scope="module")
+def cli_registered_agent(agent, agent_conf):
+    temboard(
+        "register-instance",
+        "localhost",
+        agent_conf.get("temboard", "port"),
+        "--environment",
+        "default",
+    )
+    return agent
+
+
+@pytest.fixture(scope="module")
+def browse_settings_instance(browse_instance, browser_session):
+    """Go to Settings tab of current instance."""
+    browser_session.select("#linkSettings").click()
+
+
+def test_delete_instance(
+    cli_registered_agent, browser, browse_settings_instance, agent_conf
+):
+    port = agent_conf.get("temboard", "port")
+    browser.mincount("#tableInstances tr", 2)
+    cli_instance_tr = [
+        tr
+        for tr in browser.select_all("#tableInstances tr")
+        if f"localhost:{port}" in tr.text
+    ][0]
+    cli_instance_tr.find_element(
+        by=By.CSS_SELECTOR, value="td button.buttonDelete"
+    ).click()
+    browser.clickable("#modalDeleteInstance button#buttonDelete").click()
+    # Check the instance has been deleted
+    assert all(
+        f"localhost:{port}" not in tr.text
+        for tr in browser.select_all("#tableInstances tbody tr")
+    )
 
 
 def test_about_instance(registered_agent, agent_conf, browser, ui_url):
